Finish vftint

+++ code/vftintrmh_l_s.h
@@ -1,3 +1,3 @@
 for (int i = 0; i < 2; i++) {
-  dst.dword[i] = a.fp32[i + 2]; // rounding mode is not expressed in C
+  dst.dword[i] = (s32)a.fp32[i + 2]; // rounding mode is not expressed in C
 }
