+++ code/vftintrne_w_s.h
@@ -1,3 +1,3 @@
 for (int i = 0; i < 4; i++) {
-  dst.word[i] = a.fp32[i]; // rounding mode is not expressed in C
+  dst.word[i] = (s32)a.fp32[i]; // rounding mode is not expressed in C
 }
